remove duplicated entities
