--- api/src/modules/model/entities/data-collection-and-field-costs.entity.ts
@@ -1,20 +1,0 @@
-import {
-  Entity,
-  Column,
-  PrimaryGeneratedColumn,
-  OneToMany,
-  BaseEntity,
-} from 'typeorm';
-import { BaseData } from '@shared/entities/base-data.entity';
-
-@Entity('data_collection_and_field_costs')
-export class DataCollectionAndFieldCosts extends BaseEntity {
-  @PrimaryGeneratedColumn('uuid')
-  id: string;
-
-  @Column('decimal', { name: 'field_cost' })
-  fieldCost: number;
-
-  @OneToMany(() => BaseData, (baseData) => baseData.dataCollectionAndFieldCosts)
-  baseData: BaseData[];
-}
