+++ shared/lib/db-entities.ts
@@ -1,19 +1,19 @@
-// import { User } from "@shared/entities/users/user.entity";
-// import { ApiEventsEntity } from "@api/modules/api-events/api-events.entity";
-// import { Country } from "@api/modules/model/entities/country.entity";
-// import { BaseData } from "@api/modules/model/base-data.entity";
-// import { ProjectSize } from "@api/modules/model/entities/project-size.entity";
-// import { FeasibilityAnalysis } from "@api/modules/model/entities/feasability-analysis.entity";
-// import { ConservationPlanningAndAdmin } from "@api/modules/model/entities/conservation-and-planning-admin.entity";
-// import { DataCollectionAndFieldCosts } from "@api/modules/model/entities/data-collection-and-field-costs.entity";
-//
-// export const BACKEND_DB_ENTITIES = [
-//   User,
-//   ApiEventsEntity,
-//   Country,
-//   BaseData,
-//   ProjectSize,
-//   FeasibilityAnalysis,
-//   ConservationPlanningAndAdmin,
-//   DataCollectionAndFieldCosts,
-// ];
+import { User } from "@shared/entities/users/user.entity";
+import { ApiEventsEntity } from "@api/modules/api-events/api-events.entity";
+import { Country } from "@shared/entities/country.entity";
+import { BaseData } from "@shared/entities/base-data.entity";
+import { ProjectSize } from "@shared/entities/project-size.entity";
+import { FeasibilityAnalysis } from "@shared/entities/feasability-analysis.entity";
+import { ConservationPlanningAndAdmin } from "@shared/entities/conservation-and-planning-admin.entity";
+import { DataCollectionAndFieldCosts } from "@shared/entities/data-collection-and-field-costs.entity";
+
+export const COMMON_DATABASE_ENTITIES = [
+  User,
+  ApiEventsEntity,
+  Country,
+  BaseData,
+  ProjectSize,
+  FeasibilityAnalysis,
+  ConservationPlanningAndAdmin,
+  DataCollectionAndFieldCosts,
+];
